+++ framework/model/src/main/java/is/codion/framework/model/AbstractEntityTableModel.java
@@ -218,9 +218,9 @@ public abstract class AbstractEntityTableModel<M extends EntityModel<M, E, T, R>
 	}
 
 	private void bindEvents() {
-		editModel.editor().events().afterInsert().addConsumer(this::onInsert);
-		editModel.editor().events().afterUpdate().addConsumer(this::onUpdate);
-		editModel.editor().events().afterDelete().addConsumer(this::onDelete);
+		editModel.editor().events().after().insert().addConsumer(this::onInsert);
+		editModel.editor().events().after().update().addConsumer(this::onUpdate);
+		editModel.editor().events().after().delete().addConsumer(this::onDelete);
 		editModel.editor().entity().addConsumer(this::onEntityChanged);
 		selection().item().addConsumer(editModel.editor().entity()::set);